worker: Move code of FakeWorkerForBuilder to FakeProtocolCommand
