--- worker/buildbot_worker/test/fake/workerforbuilder.py
@@ -1,71 +1,0 @@
-# This file is part of Buildbot.  Buildbot is free software: you can
-# redistribute it and/or modify it under the terms of the GNU General Public
-# License as published by the Free Software Foundation, version 2.
-#
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-# details.
-#
-# You should have received a copy of the GNU General Public License along with
-# this program; if not, write to the Free Software Foundation, Inc., 51
-# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-#
-# Copyright Buildbot Team Members
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import pprint
-
-
-class FakeWorkerForBuilder(object):
-
-    """
-    Simulates a WorkerForBuilder, but just records the updates from sendUpdate
-    in its updates attribute.  Call show() to get a pretty-printed string
-    showing the updates.  Set debug to True to show updates as they happen.
-    """
-    debug = False
-
-    def __init__(self, basedir="/workerbuilder/basedir"):
-        self.updates = []
-        self.basedir = basedir
-        self.unicode_encoding = 'utf-8'
-
-    def sendUpdate(self, data):
-        if self.debug:
-            print("FakeWorkerForBuilder.sendUpdate", data)
-        self.updates.append(data)
-
-    def show(self):
-        return pprint.pformat(self.updates)
-
-    # Returns a Deferred
-    def protocol_update_upload_file_close(self, writer):
-        return writer.callRemote("close")
-
-    # Returns a Deferred
-    def protocol_update_upload_file_utime(self, writer, access_time, modified_time):
-        return writer.callRemote("utime", (access_time, modified_time))
-
-    # Returns a Deferred
-    def protocol_update_upload_file_write(self, writer, data):
-        return writer.callRemote('write', data)
-
-    # Returns a Deferred
-    def protocol_update_upload_directory(self, writer):
-        return writer.callRemote("unpack")
-
-    # Returns a Deferred
-    def protocol_update_upload_directory_write(self, writer, data):
-        return writer.callRemote('write', data)
-
-    # Returns a Deferred
-    def protocol_update_read_file_close(self, reader):
-        return reader.callRemote('close')
-
-    # Returns a Deferred
-    def protocol_update_read_file(self, reader, length):
-        return reader.callRemote('read', length)
