+++ worker/buildbot_worker/test/unit/test_commands_shell.py
@@ -50,6 +50,6 @@ class TestWorkerShellCommand(CommandTestMixin, unittest.TestCase):
         # note that WorkerShellCommand does not add any extra updates of it own
         self.assertUpdates(
             [{'hdr': 'headers'}, {'stdout': 'hello\n'}, {'rc': 0}],
-            self.builder.show())
+            self.protocol_command.show())
 
     # TODO: test all functionality that WorkerShellCommand adds atop RunProcess
